requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
cloudscraper==1.2.71
idna==3.6
mysql-connector-python==8.2.0
protobuf==4.21.12
pyparsing==3.1.1
requests==2.31.0
requests-toolbelt==1.0.0
soupsieve==2.5
urllib3==2.1.0